% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_sample_comparison.R
\name{beta_sample_comparison}
\alias{beta_sample_comparison}
\title{Sampling beta distributions and compute the probability of theta1 < theta2}
\usage{
beta_sample_comparison(sample_size, beta_a1, beta_b1, beta_a2, beta_b2)
}
\arguments{
\item{sample_size}{how many samples to produce}

\item{beta_a1}{first parameter of theta1}

\item{beta_b1}{second parameter of theta1}

\item{beta_a2}{first parameter of theta2}

\item{beta_b2}{second parameter of theta2}
}
\value{
the probability of theta1 < theta2
}
\description{
Sampling beta distributions and compute the probability of theta1 < theta2
}
\examples{
beta_sample_comparison(5000, 58, 44, 31, 21)
}
