% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_poisson_ppd.R
\name{gamma_poisson_ppd}
\alias{gamma_poisson_ppd}
\title{Title Poisson model of parameter theta, given gamma prior and the data. Sampling the posterior predictive distribution.}
\usage{
gamma_poisson_ppd(sample_size = sample_size, gamma_a1 = gamma_a1,
  gamma_b1 = gamma_b1, y1 = y1, gamma_a2 = gamma_a2,
  gamma_b2 = gamma_b2, y2 = y2, using_MCMC = FALSE, plot = FALSE)
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{gamma_a1}{First parameter of theta1.}

\item{gamma_b1}{Second parameter of theta1.}

\item{y1}{Given data of the Possion distribution with gamma prior for the parameter theta1.}

\item{gamma_a2}{First parameter of theta2.}

\item{gamma_b2}{Second parameter of theta2.}

\item{y2}{Given data of the Possion distribution with gamma prior for the parameter theta2.}

\item{using_MCMC}{Whether to use MCMC to sampling the posterior predictive distribution.}

\item{plot}{whether to plot the parameter samplings.}
}
\value{
Sampling of the posterior predictive distribution.
}
\description{
Title Poisson model of parameter theta, given gamma prior and the data. Sampling the posterior predictive distribution.
}
\examples{
gamma_poisson_ppd(sample_size = 5000, gamma_a1 = 2, gamma_b1 = 1, y1 = c(1,0,0,1,2,2,1,5,2,0,0,0,0,0,0,1,1,1,0,0,0,1,1,2,1,3,2,0,0,3,0,0,0,2,1,0,2,1,0,0,1,3,0,1,1,0,2,0,0,2,2,1,3,0,0,0,1,1), gamma_a2 = 2, gamma_b2 = 1, y2 = c(2,2,1,1,2,2,1,2,1,0,2,1,1,2,0,2,2,0,2,1,0,0,3,6,1,6,4,0,3,2,0,1,0,0,0,3,0,0,0,0,0,1,0,4,2,1,0,0,1,0,3,2,5,0,1,1,2,1,2,1,2,0,0,0,2,1,0,2,0,2,4,1,1,1,2,0,1,1,1,1,0,2,3,2,0,2,1,3,1,3,2,2,3,2,0,0,0,1,0,0,0,1,2,0,3,3,0,1,2,2,2,0,6,0,0,0,2,0,1,1,1,3,3,2,1,1,0,1,0,0,2,0,2,0,1,0,2,0,0,2,2,4,1,2,3,2,0,0,0,1,0,0,1,5,2,1,3,2,0,2,1,1,3,0,5,0,0,2,4,3,4,0,0,0,0,0,0,2,2,0,0,2,0,0,1,1,0,2,1,3,3,2,2,0,0,2,3,2,4,3,3,4,0,3,0,1,0,1,2,3,4,1,2,6,2,1,2,2), using_MCMC = FALSE, plot = TRUE)
}
