% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_effective.R
\name{gibbs_effective}
\alias{gibbs_effective}
\title{Title Analysis of the mixing efficiency of Gibbs sampling.}
\usage{
gibbs_effective(sample_size, x, y, tau_beta_square, tau_c_square,
  confidence_interval = c(0.025, 0.975))
}
\arguments{
\item{sample_size}{The sample size used in analyzing the mixing of Gibbs sampling.}

\item{x}{The data set need to be analyzed.}

\item{y}{The indicator vector of the data x.}

\item{tau_beta_square}{Variance of the normal distribution of beta.}

\item{tau_c_square}{Variance of the normal distribution of c.}

\item{confidence_interval}{Confidence interval to be calculated.}
}
\value{
Confidence interval of beta, and the probability of beta being positive given \eqn{\boldsymbol{x}} and \eqn{\boldsymbol{y}}.
}
\description{
Title Analysis of the mixing efficiency of Gibbs sampling.
}
\examples{
gibbs_effective(sample_size = 30000, x = divorce$V1, y = divorce$V2, 
tau_beta_square = 16, tau_c_square = 16)
}
