% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdingFunctions.R
\name{hard}
\alias{hard}
\title{Hard-thresholding function}
\usage{
hard(a, lambda)
}
\arguments{
\item{a}{scalar to be tresholded}

\item{lambda}{threshold value, should be non-negative}
}
\value{
Hard-tresholded value of a
}
\description{
Hard-thresholding function
}
\examples{
hard(3,1)
}
