% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_sample_comparison.R
\name{posterior_predictive_sample_comparison}
\alias{posterior_predictive_sample_comparison}
\title{Poisson model with gamma prior, compare the two variables}
\usage{
posterior_predictive_sample_comparison(sample_size = sample_size,
  gamma_a1 = gamma_a1, gamma_b1 = gamma_b1, gamma_a2 = gamma_a2,
  gamma_b2 = gamma_b2, theta1_smaller_than_theta2 = TRUE,
  plot = FALSE)
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{gamma_a1}{First parameter of theta1.}

\item{gamma_b1}{Second parameter of theta1.}

\item{gamma_a2}{First parameter of theta2.}

\item{gamma_b2}{Second parameter of theta2.}

\item{theta1_smaller_than_theta2}{How to make the comparison. If FALSE, compute theta1 > theta2.}

\item{plot}{Whether to plot the figure showing the tendency of probability change.}
}
\value{
The probability vector or value
}
\description{
Poisson model with gamma prior, compare the two variables
}
\examples{
posterior_predictive_sample_comparison(sample_size = 5000, beta_a1 = 12*(1:100)+113, beta_b1 = 13+(1:100), beta_a2 = 237, beta_b2 = 20, theta1_smaller_than_theta2 = FALSE, plot = TRUE)
}
