% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Title Analyze the effect of prior data on the model}
\usage{
sensitivity_analysis(sample_size, y_bar.1, standard_deviation.1, y_bar.2,
  standard_deviation.2, mu_0, sigma_0_square, n, kappa_0, nu_0,
  theta.1_smaller_than_theta.2 = TRUE, plot = TRUE)
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{y_bar.1}{Mean value of the first group of data.}

\item{standard_deviation.1}{Standard deviation of the first group of data.}

\item{y_bar.2}{Mean value of the second group of data.}

\item{standard_deviation.2}{Standard deviation of the second group of data.}

\item{nu_0}{}
}
\description{
Title Analyze the effect of prior data on the model
}
\examples{
sensitivity_analysis(sample_size = 10000, y_bar.1 = 75.2, standard_deviation.1 = 7.3, y_bar.2 = 77.5, standard_deviation.2 = 8.1, mu_0 = 75, sigma_0_square = 100, n = 16, kappa_0 = c(1,2,4,8,16,32), nu_0 = c(1,2,4,8,16,32)) # sigma_0_square = 4?
}
