% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdingFunctions.R
\name{soft}
\alias{soft}
\title{Soft-tresholding function}
\usage{
soft(a, lambda)
}
\arguments{
\item{a}{scalar to be tresholded}

\item{lambda}{threshold value, should be non-negative}
}
\value{
Soft-tresholded value of a
}
\description{
Soft-tresholding function
}
\examples{
soft(3, 1)
}
