% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_prior_sensitivity.R
\name{two_prior_sensitivity}
\alias{two_prior_sensitivity}
\title{Title Analysis the sensitivity of changing the parameters of Gamma distribution, based on the dependency of two Poisson distributions parameters.}
\usage{
two_prior_sensitivity(sample_size, gamma_theta_a, gamma_theta_b, y.1, y.2,
  gamma_gamma_a, gamma_gamma_b)
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{gamma_theta_a}{The first parameter of Gamma distribution of \eqn{\theta}.}

\item{gamma_theta_b}{The second parameter of Gamma distribution of \eqn{\theta}.}

\item{y.1}{The first group of data.}

\item{y.2}{The second group of data.}

\item{gamma_gamma_a}{The first parameter of Gamma distribution of \eqn{\gamma}.}

\item{gamma_gamma_b}{The second parameter of Gamma distribution of \eqn{\gamma}.}
}
\value{
The expectation of the difference between parameters of Poisson distributions given the data.
}
\description{
Title Analysis the sensitivity of changing the parameters of Gamma distribution, based on the dependency of two Poisson distributions parameters.
}
\examples{
two_prior_sensitivity(sample_size = 5000, gamma_theta_a = 2, gamma_theta_b = 1,
y.1 = menchild30bach, y.2 = menchild30nobach,
gamma_gamma_a = 2^(seq(3, 7)), gamma_gamma_b = 2^(seq(3, 7)))
}
