% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_check.R
\name{vector_check}
\alias{vector_check}
\title{Title Check if the object is a vector or not}
\usage{
vector_check(object, length = NULL)
}
\arguments{
\item{object}{Vector to be tested.}

\item{length}{Expected length.}
}
\value{
True or False. Vector of length 1 is not treated as a vector, thus returning false, unless the specified length is 1.
}
\description{
Title Check if the object is a vector or not
}
\examples{
vector_check(object = c(1)) == FALSE
vector_check(object = c(1), length = 1) == TRUE
vector_check(object = c(1, 1)) == TRUE
vector_check(object = c(1, 2), length = 2) == TRUE
vector_check(object = c(1, 1), length = 1) == FALSE
}
