% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_poisson_ppd.R
\name{gamma_poisson_ppd}
\alias{gamma_poisson_ppd}
\title{Title Poisson model of parameter theta, given gamma prior and the data. Sampling the posterior predictive distribution.}
\usage{
gamma_poisson_ppd(sample_size, gamma_a1, gamma_b1, y1, gamma_a2, gamma_b2,
  y2, confidence_interval = NULL, using_MCMC = FALSE,
  poisson_fitting_mean = NULL)
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{gamma_a1}{First parameter of theta1.}

\item{gamma_b1}{Second parameter of theta1.}

\item{y1}{Given data of the Possion distribution with gamma prior for the parameter theta1.}

\item{gamma_a2}{First parameter of theta2.}

\item{gamma_b2}{Second parameter of theta2.}

\item{y2}{Given data of the Possion distribution with gamma prior for the parameter theta2.}

\item{confidence_interval}{Confidence interval to be calculated.}

\item{using_MCMC}{Whether to use MCMC to sampling the posterior predictive distribution.}

\item{poisson_fitting_mean}{Some given parameter for Poisson distribution to be tested.}
}
\value{
Sampling of the posterior predictive distribution.
}
\description{
Title Poisson model of parameter theta, given gamma prior and the data. Sampling the posterior predictive distribution.
}
\examples{
gamma_poisson_ppd(sample_size = 5000, gamma_a1 = 2, gamma_b1 = 1, 
y1 = menchild30bach, gamma_a2 = 2, gamma_b2 = 1, 
y2 = menchild30nobach, using_MCMC = FALSE)

gamma_poisson_ppd(sample_size = 5000, gamma_a1 = 2, gamma_b1 = 1, 
y1 = menchild30bach, gamma_a2 = 2, gamma_b2 = 1, 
y2 = menchild30nobach, using_MCMC = TRUE)

gamma_poisson_ppd(sample_size = 5000, gamma_a1 = 2, gamma_b1 = 1, 
y1 = menchild30bach, gamma_a2 = 2, gamma_b2 = 1, 
y2 = menchild30nobach, confidence_interval = c(0.025, 0.975), 
using_MCMC = FALSE, poisson_fitting_mean = 1.4)
}
