% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_conjugate_family.R
\name{normal_gamma_conjugate_family}
\alias{normal_gamma_conjugate_family}
\title{Title Univariate normal distribution with normal-gamma conjugate distribution}
\usage{
normal_gamma_conjugate_family(sample_size, mu_0, sigma_0_sequare, kappa_0,
  nu_0, y.1, y.2 = NULL, y.3 = NULL, confidence_interval = c(0.025,
  0.975))
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{mu_0}{The mean of normal distribution of theta given sigma squared.}

\item{sigma_0_sequare}{Sample variance.}

\item{kappa_0}{The number of prior observations.}

\item{nu_0}{The number of samples of prior observations.}

\item{y.1}{Data set number 1.}

\item{y.2}{Data set number 2.}

\item{y.3}{Data set number 3.}

\item{confidence_interval}{Confidence interval to be calculated.}
}
\value{
The comparison of three data sets.
}
\description{
Title Univariate normal distribution with normal-gamma conjugate distribution
}
\examples{
normal_gamma_conjugate_family(sample_size = 10000, mu_0 = 5, sigma_0_sequare = 4, 
kappa_0 = 1, nu_0 = 2, y.1 = school1, 
y.2 = school2, y.3 = school3, confidence_interval = c(0.025, 0.975))
}
