% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_sample_comparison.R
\name{posterior_predictive_sample_comparison}
\alias{posterior_predictive_sample_comparison}
\title{Poisson model with gamma prior, compare the two parameters}
\usage{
posterior_predictive_sample_comparison(sample_size = sample_size,
  gamma_a1 = gamma_a1, gamma_b1 = gamma_b1, gamma_a2 = gamma_a2,
  gamma_b2 = gamma_b2, tildeya_smaller_than_tildeyb = TRUE,
  plot = FALSE)
}
\arguments{
\item{sample_size}{How many samples to produce.}

\item{gamma_a1}{First parameter of theta1.}

\item{gamma_b1}{Second parameter of theta1.}

\item{gamma_a2}{First parameter of theta2.}

\item{gamma_b2}{Second parameter of theta2.}

\item{tildeya_smaller_than_tildeyb}{How to make the comparison. If FALSE, compute theta1 > theta2.}

\item{plot}{Whether to plot the figure showing the tendency of probability change.}
}
\value{
The probability vector or value
}
\description{
Poisson model with gamma prior, compare the two parameters
}
\examples{
posterior_predictive_sample_comparison(sample_size = 5000, gamma_a1 = 237, 
gamma_b1 = 20, gamma_a2 = 12*(1:100)+113, gamma_b2 = 13+(1:100), 
tildeya_smaller_than_tildeyb = FALSE, plot = TRUE)
}
